- Move helper functions out of __main__.py into helpers.py - added unittests for helpers.py
